版本更新

+++ rent-admin/src/main/java/com/xxz/rent/dto/SmsFlashPromotionSessionDetail.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class SmsFlashPromotionSessionDetail extends SmsFlashPromotionSession {
     @Setter
     @Getter
-    private Integer productCount;
+    private Long productCount;
 }
